fix: .env 파일 수정
