+++ chap03/index.html
@@ -73,7 +73,6 @@
       sizes="16x16"
       href="./public/faviconFiles/favicon-16x16.png"
     />
-    <link rel="manifest" href="./public/faviconFiles/manifest.json" />
     <meta name="msapplication-TileColor" content="#ffffff" />
     <meta
       name="msapplication-TileImage"
